requirements.txt
dnspython
dnslib
requests
requests-toolbelt